services/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = ""
authors = [
    {name = "Chetanand Meher",email = "chetanandmeher1282004@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "kafka-python (>=2.3.1,<3.0.0)",
    "pydantic-settings (>=2.14.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
